% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_M.R
\name{calc_M}
\alias{calc_M}
\title{Calculate M}
\usage{
calc_M(d)
}
\arguments{
\item{d}{vector of points in the series}
}
\description{
Determines the maximum value of M, where 2^M < N (length of original series)
}
\author{
W. S. Drysdale
}
