% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_vec.R
\name{group_vec}
\alias{group_vec}
\title{Group Vector}
\usage{
group_vec(vec, n)
}
\arguments{
\item{vec}{vector}

\item{n}{length of groups}
}
\description{
Gathers sequential items in a vector (e.g file names) into a list of vectors, of length n
}
\author{
W. S. Drysdale
}
