% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_2M.R
\name{map_to_2M}
\alias{map_to_2M}
\title{Map to 2^M}
\usage{
map_to_2M(d, time_res = 0.2)
}
\arguments{
\item{d}{vector of points in the series}

\item{time_res}{time resolution of original series in seconds}
}
\description{
Multiresolution Decomposition requires the length of the series to be 2^M points, where M is an interger > 0. \cr
For the majority of series this is not the case. map_to_2M interpolates the series on the nearest 2^M length series, where 2^M < N.
}
\author{
W. S. Drysdale
}
