% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrd_summarise.R
\name{mrd_and_summarise}
\alias{mrd_and_summarise}
\title{MRD and Summarise}
\usage{
mrd_and_summarise(dat, id = "1hour", col1 = "w_met",
  col2 = c("FD_mole_NO1", "FD_mole_NO2", "T_air", "uv_met"),
  group_names = NULL, time_res = 0.2, inc_w_n = F)
}
\arguments{
\item{dat}{List of data.frames to map \code{multi_res_decomp} over}

\item{id}{string identifying this run - so that results can be bound and compared}

\item{col1}{name or vector of first param. if this is shorter than col2, the first value will be repeated}

\item{col2}{name or vector of second param}

\item{group_names}{optional, what should each output be named. Will be ignored if it is of different length to col2}
}
\description{
Runs the Multiresolution decompostion over multiple files and returns the mean of all decompositions \cr
Currently only implements cospectral case
}
\author{
W. S. Drysdale
}
