% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_segs.R
\name{n_segs}
\alias{n_segs}
\title{n_segs}
\usage{
n_segs(d, i)
}
\arguments{
\item{d}{vector of data}

\item{i}{2^i is the number of segments returned}
}
\value{
list of segments
}
\description{
Split the series into the correct number of segments for this iteration of the MRD
}
\author{
W. S. Drysdale
}
