% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ncsv.R
\name{read_ncsv}
\alias{read_ncsv}
\title{Read n-csv}
\usage{
read_ncsv(ncsv, ...)
}
\arguments{
\item{ncsv}{vector of filenames}
}
\description{
Shorthand for \code{map_df(filenames,read.csv,...)} so it can easily be passed to \code map
}
\author{
W. S. Drysdale
}
