% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smam.R
\name{smam}
\alias{smam}
\title{Second Moment about the Mean}
\usage{
smam(w_n, M, m, psi_n = NULL)
}
\arguments{
\item{w_n}{vector of points from current stage of decompostion}

\item{M}{Maximum value of m, 2^M is the length of the original series}

\item{m}{is the current value of m for this iteration}

\item{psi_n}{second vector of points for cospectral case}
}
\description{
Calculate the second moment about the mean for a vector of points (or the cospectral equivalent if psi_n is supplied)
}
\author{
W. S. Drysdale
}
